+++ src/EliteChroma.Core/Layers/InterfaceModeLayer.cs
@@ -21,6 +21,17 @@ namespace EliteChroma.Core.Layers
             CustomKeyboardEffect k = canvas.Keyboard;
             k[Key.Escape] = Colors.InterfaceMode;
 
+            if (Game.IsWalking)
+            {
+                return;
+            }
+
+            if (Game.Status.HasFlag(Flags2.InTaxi))
+            {
+                ApplyColorToBinding(canvas.Keyboard, InterfaceMode.Select, Colors.InterfaceMode);
+                return;
+            }
+
             ApplyColorToBinding(canvas.Keyboard, InterfaceMode.All, Colors.InterfaceMode);
         }
     }