+++ CHANGELOG.md
@@ -4,7 +4,7 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 
-## [Unreleased](https://github.com/poveden/EliteChroma/compare/v1.13.1...HEAD)
+## [1.14.0](https://github.com/poveden/EliteChroma/compare/v1.13.1...v1.14.0) — 2021-08-19
 
 ### Fixed